safer default configuration value (#6)
This changes the default deletion behaviour from immediate deletion to
scheduled deletion after 24h. Users should not be surprised with
accidental data loss if they run the controller for the first time, but
rather be able to cancel unwanted deletion requests before they are
executed.

This also sets the statefulset label selector to a sane default so that
users have to explicitly set it to an empty string if they really want
the controller to manage all statefulsets that are around.